requirements.txt
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dataclasses-json==0.5.6
dateutils==0.6.12
decorator==5.1.0
EasyProcess==0.3
facepy==1.0.12
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.3.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib==1.0.4
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
lxml==4.8.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.2.0
outcome==1.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
pika-stubs==0.1.3
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pymongo==4.0.1
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-facebook-api==0.13.0
pytz==2021.3
PyVirtualDisplay==2.2
requests==2.26.0
requests-oauthlib==1.3.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
